fix(desktop): use f64 for timestamps to match JS backend floats
Amp-Thread-ID: https://ampcode.com/threads/T-019fa07b-6855-70b9-86f9-f22cf76f63ee

+++ apps/desktop/src-tauri/src/models/api.rs
@@ -69,9 +69,9 @@ pub struct ProjectInfo {
     pub id: Option<String>,
     pub name: String,
     pub path: String,
-    pub created_at: Option<u64>,
-    pub updated_at: Option<u64>,
-    pub last_modified: Option<u64>,
+    pub created_at: Option<f64>,
+    pub updated_at: Option<f64>,
+    pub last_modified: Option<f64>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
